+++ Editor/Search/Internal/PropertiesSearchBackend.cs
@@ -11,21 +11,47 @@ namespace Unity.Properties.UI.Internal
     {
         class SearchQuery : ISearchQuery<TData>
         {
+            static readonly string[] k_SupportedOperators = FilterOperator.GetSupportedOperators<TData>();
+
             readonly Func<TData, IEnumerable<string>> m_GetSearchDataFunc;
+            readonly string[] m_Tokens;
 
             public string SearchString { get; }
 
-            public ICollection<string> Tokens => SearchString.Split(' ');
-            
+            public ICollection<string> Tokens => m_Tokens;
+
             public SearchQuery(string searchString, Func<TData, IEnumerable<string>> getSearchDataFunc)
             {
                 SearchString = searchString;
+                m_Tokens = string.IsNullOrWhiteSpace(SearchString) ? Array.Empty<string>() : SearchString.Split(' ');
                 m_GetSearchDataFunc = getSearchDataFunc;
             }
             
             public IEnumerable<TData> Apply(IEnumerable<TData> data)
             {
-                return data.Where(d => m_GetSearchDataFunc(d).Any(s => s.IndexOf(SearchString, StringComparison.OrdinalIgnoreCase) >= 0));
+                var tokens = FilteredSupportedTokens().ToArray();
+                return tokens.Length == 0
+                    ? data
+                    : data.Where(d => m_GetSearchDataFunc(d).Any(s => tokens.Any(t => s.IndexOf(t, StringComparison.OrdinalIgnoreCase) >= 0)));
+            }
+
+            IEnumerable<string> FilteredSupportedTokens()
+            {
+                foreach (var token in m_Tokens)
+                {
+                    var isTokenSupported = true;
+                    foreach (var supportedOperator in k_SupportedOperators)
+                    {
+                        if (token.IndexOf(supportedOperator, StringComparison.OrdinalIgnoreCase) < 0)
+                            continue;
+
+                        isTokenSupported = false;
+                        break;
+                    }
+
+                    if (isTokenSupported)
+                        yield return token;
+                }
             }
         }